Make the member resolver one authority, because a second copy decides
StoreSqlClockDisciplineTests carried a byte-identical copy of the
pre-fix regex, and its own doc comment said what it was for: "the waiver
key and the failure message". It feeds the resolved name into
Waived.Contains(file + ":" + member), so a wrong name there is a wrong
DECISION, not a wrong message — it can miss a legitimate waiver, or
collide with an unrelated waived key and swallow a real bare-clock
finding. Fixing the message-only copy and filing that one is shipping
the inert half.

The resolver moves to CSharpMemberMap and both pins use it, the
treatment CSharpSourceWalker got in #2913 and RepoFile in #3090.
Correcting the copy in place instead would leave two implementations
that have already been shown to diverge.

Measured before moving: 21 of the 140 SQL literals in the store tree
change label — if, using, while, Select, NpgsqlCommand — and none of
them is the one waived key, which resolves to RefreshSql under both. A
resolver that stops resolving now reds that suite by name, which is how
the wiring is shown to be live rather than merely compiled.
